cppcheck: fix comparison of pointers pointed to different objects

+++ kernel/rom_dev.c
@@ -26,7 +26,7 @@ ssize_t rom_dev_read(struct file *filp, char *buf, size_t size, loff_t offset)
 {
     char *read_addr = &_section_rom_start + offset;
 
-    if((read_addr + size) > &_section_rom_end) {
+    if((uint32_t)(read_addr + size) > (uint32_t)&_section_rom_end) {
         return EFAULT;
     }
